Added security routing to two new pages
Added the security process to reroute any attempt to access userpage or
admin page - rerouted to the loginpage. The login page will be changed
for a form that once the password check (conducted by spring frameowrk
and mysql) has been complete - access will be granted.
added loginpage, userpage and adminpage