Added security login form
added security login form - links to database via springs security
(j_spring_security_check) - checks with database to see if user is in
table - for both admin and a user.
Added login buttons for startpage to assist with navigation to admin or
userpages.
Need to add secuity logout for both admin and any users.