Added control over creation & deletion of Admins
Ensured that one admin has to remain on the server in order to delete
other Admins. (errordeleteadminuser page).
Changed the appearance of the delete admin page with warnings, this is
also that same for deleting a user.
Changed how users are added - now add an admin or a user.
Removed the possibility of creating an Admin which would automatically
create a user. This is no longer possible. No Admin will have a
user_role.
Removed unnecessary code from adminpage/allusers - this is all done via
sql - thus java code no longer required.
Added back pages to deleteuser pages to allow a level of fluid control.
Added a box to the add user page to tidy it up  - need to consider
adding another row entry for confirming password - need to look how to
do this.