Added Upload Page, and Avatar practice
Added a basic upload file page - allows the upload of multiple files.
Uses the MultipartFile class - which assist with the file transfer.
At present files can be uploaded, no means to test whether the file is
already there. File will not be copied across if already there but will
say it was successful.

Added a few extra fancy buttons - practicing with xml/css

Added a static image to display in edit profile page - this will become
the Avatar users can have for the page - simply aesthetic.