Updated the file upload page and control
Updated the pages for uploading files.

Controlled the matters surrounding not selecting a file to upload, if
the file you are uploading already exists (does not (upload), and shows
when you selected file to upload that are not present with those that
are present.

Pages layouts updated.

TO DO: Add a file Manager central page to assist with accessing public
private folders, deletion etc.