Amended back buttons to upload page
Upload page is used for both public and private - at this stage only
programmed for public. assists with DRY...

Correctly update the uploadfile page when calling itself when no file
has been selected - this allows the links that are present to be used
correctly - to include the public folder comment and the insertion of
the user

Controlled access with intercept to public folders. - ignored the url of
the public page to allow this file to be accessed by anyone - warning to
be added. Could have a second login in page using a simple username and
password - hardcoded.