Added the ability to download files
using two websites as a reference -
 references http://docs.spring.io/spring/docs/1.2.x/api/org/springframework/util/FileCopyUtils.html
 & http://owlsayswoot.therandomist.com/2011/12/12/how-to-download-files-with-springmvc/ 15/07/2015)

able to identify the file type - and correctly check if file existed - if so download.
Using HttpServletResponse to get the file type, set the header, and length then using
FileCopyUtils.copy to allow the file to be outputted.

on clicking download the file in chrome is automatically downloaded to user folder downloads. opened correctly