Reshuffled Security intercept handlers & added SecurityChecker class
the intercepts stopped working for some reason so reshuffled and are now working

Added a class SecurityChecker.java - this has two static functions 1. checks is the folderType is public or private. 2. checks if the username is the same as the user who has logged into the session.

Each method that has a generic { } in its url has had these checks added - if i duplicated code this would nt be necessary. This still saves lots of coding.
Username is checked first - if fails then send back to homePage - when they click on user login - they will login straight to their profile page as they have logged in correctly. Avoid them seeing other peoples folders.

The same for folderType - this check sends the user back to their profile page if the test is failed. Stops them typing in any old name and returning that page.