Added buttons to the Layout and changed several pages
Added buttons to the header in a row,
Added  the username and role of the user to the footer.

A constant logout button in far right of header.

usereditpage, profile page, login page, start page and user folder page
have now been changed.

user folder page needs to have the files listed as hyper links and a new
page needs to be created to offer the same options on each file