Added additional security to Admin alluser pages and users attempting to access other users pages.
Admin all users page - added reset password to each user, not admin. it redirect to a new page with the new password. at present it resets to name@123. could make this random.
In order for admin to carryout this process, the current logged in user is check as to whether they are an admin - new method added in securitychecker - isAdmin(User user). if they are not an admin then the password reset fails. additionally the individual having their password reset is checked if they are on the system,

Added additional security for users trying to access other user profile page- instead of redirecting back to main page but remaining logged in - now the system logs them out.