Amended the file copying process and feedback
If the file exists in the destination directory it is not copied and the appropriate warning is given on the userfoldepage.
If the file failed - an error is shown and the user is asked to retry
If the file has been copied then a confirmation is given in the message folder.

The controller method in fileController checks whether the file is already in the public folder before attempting to copy.