Added a method to identify image files. Added the function to view an image file in file handling page.
Added a method in UserFileControl - to identify if the string file name provided is an image file - iterates through a list and compares.

Updated the filehandling page - if a file is of the file image type as specified then a new button appears that lets you view that said image. if however the file is not an image then it cannot be viewed, and the button does not appear.

Imageviewer jsp is called to view the image - it now has a back button.

Tested and working well