Added Password Hashing for security
Added the BCryptPasswordEncoder class to assist with hashing the
password. old password no longer worked and thus the database stores the
hash not the actual string password.

the encoding takes place one a user is created by the admin - at this
stage his password that was used for him is encoded and the user object
gets set the encoded password.

added the password encoder to the security.xml

removed a few unnecessary files