Added a restful practice greeting
Following - http://spring.io/guides/gs/rest-service/, i added the
restful return of json information on the screen - this was a practice
to see if the service would allow a restful response to a JSON request
per se.

This is in preparation for android linking to the service.

Added new dependencies - jackson.core (json), project lobok - helps with
boiler plate code, and spring release repositorys.  marked up a s
android in the POM.XML

finally added spring boot framework - plugin for maven again in pom.xml

the practice classes were greeting and greeting controller called
application.

these will now be changed to accomodate the user and the data from the
sql server