Removed old gateway login and added a new secure login
Removed old login message where the username and password were passed in
the parameters.

The new login is sent from android  - a header as part of the
HttpAuthentication class.  this sends the username and password which is
check via spring security -against the database. The password is
therefore automatically check.
This also allows spring to capture who the logged in user is and thus
more secure - carryout security checks.

If the user exists then the user is obtained from database and sent back
to application.