Added a string response to the file upload
Added string response - must be null, false or true. Thus if file
uploaded (not checked to see if the file is present in directory yet)
then return true.

login now checks for user not being null -

need to add login check for when sending data as there appears to be no
logout using android.