Added default gateway for all users who log in - security
By placing in the security.xml - "always use default url" for spring
security (once successfully logged in) then i can use the default url
"/startpage" to identify if they are an admin or user.
Once identified i can send them to the first page of users or admin
pages. This will stop people jumping ahead, but mainly i can use this
page to control failed log in attempts and remove the maxed out login
attempts from the database (yet to be implemented) . therefore the max
count of logins will be set back to 0.  table needs to be created or
column added