Added restfulcontroller from test phase
Added a restful controller class that implements the connection from the
android device.
This is taken from the initial test.class which has now been made
redundant.

The restful controller for receiving the file now takes a direction
within the url for the username in which to place the file. All files
are now placed in private files.

Tested on android device - correctly uploads to system.